Update PhoneInputNumberType.php
Update Description
A change has been made to the PhoneInputNumberType class due to the implementation of PHP 8.1 Enum features in our library. This change was necessary because the libphonenumber\PhoneNumberFormat class now uses an enum structure instead of constants.
Change Details

The toLibPhoneNumberFormat() method has been updated to return an integer value directly instead of an enum object
We now use the enum's value property to convert to the integer format
This change resolves the type mismatch error ("Return value must be of type int")

Affected Files

Ysfkaya\FilamentPhoneInput\PhoneInputNumberType.php

This update ensures compatibility with the latest version of the libphonenumber library.